Various relocation and typo fix
This poorly curated commit attempts to fix the unknown error where room_lcz fails to be built.
This fix include fixing typo in script_component, and relocating CfgSounds audios from room_lcz to anomaly module.
