+++ README.md
@@ -97,7 +97,7 @@ You can use make to manage versioning and packing up your release version.
 
 Fresh build and package it up for a release:
 ```bash
-tools/make force checkexternal release 0.5.0
+tools/make force checkexternal release 0.5.9
 ```
 
 Build just a binarized version:
@@ -117,9 +117,9 @@ You can also manage versioning through the make tool. To do this, navigate to th
 ```
 
 <<<<<<< Updated upstream
-Modify the numbers in here to represent your build version. The example listed above would be: `0.5.0.24`. This version will be set in each pbo during binarizing. It will also be used in the signature file names, along the commit hash. This will make it easier to identify the exact version that is being used.
+Modify the numbers in here to represent your build version. The example listed above would be: `0.5.9.45`. This version will be set in each pbo during binarizing. It will also be used in the signature file names, along the commit hash. This will make it easier to identify the exact version that is being used.
 =======
-Modify the numbers in here to represent your build version. The example listed above would be: `0.5.0.24`. This version will be set in each pbo during binarizing. It will also be used in the signature file names, along the commit hash. This will make it easier to identify the exact version that is being used.
+Modify the numbers in here to represent your build version. The example listed above would be: `0.5.9.45`. This version will be set in each pbo during binarizing. It will also be used in the signature file names, along the commit hash. This will make it easier to identify the exact version that is being used.
 >>>>>>> Stashed changes
