+++ addons/anomaly/CfgSounds.hpp
@@ -4,4 +4,29 @@
 
 class CfgSounds
 {
+	sounds[] = {};
+	class SCP_CB_1162_retrieve_1 {
+        sound[] = {QPATHTOF(data\SCP1162\Exchange1.ogg), 1, 1};
+        titles[] = {};
+    };
+
+    class SCP_CB_1162_retrieve_2 {
+        sound[] = {QPATHTOF(data\SCP1162\Exchange2.ogg), 1, 1};
+        titles[] = {};
+    };
+
+    class SCP_CB_1162_retrieve_4 {
+        sound[] = {QPATHTOF(data\SCP1162\Exchange4.ogg), 1, 1};
+        titles[] = {};
+    };
+
+    class SCP_CB_1162_retrieve_body_1 {
+        sound[] = {QPATHTOF(data\SCP1162\BodyHorrorExchange1.ogg), 1, 1};
+        titles[] = {};
+    };
+
+    class SCP_CB_1162_retrieve_body_3 {
+        sound[] = {QPATHTOF(data\SCP1162\BodyHorrorExchange3.ogg), 1, 1};
+        titles[] = {};
+    };
 };
