+++ addons/main/script_version.hpp
@@ -1,4 +1,4 @@
 #define MAJOR 0
 #define MINOR 5
-#define PATCHLVL 0
-#define BUILD 24
+#define PATCHLVL 9
+#define BUILD 45
