+++ addons/textures/script_component.hpp
@@ -7,12 +7,12 @@
 // #define CBA_DEBUG_SYNCHRONOUS
 // #define ENABLE_PERFORMANCE_COUNTERS
 
-#ifdef DEBUG_ENABLED_MAIN
+#ifdef DEBUG_ENABLED_TEXTURES
     #define DEBUG_MODE_FULL
 #endif
 
-#ifdef DEBUG_SETTINGS_MAIN
-    #define DEBUG_SETTINGS DEBUG_SETTINGS_MAIN
+#ifdef DEBUG_SETTINGS_TEXTURES
+    #define DEBUG_SETTINGS DEBUG_SETTINGS_TEXTURES
 #endif
 
 #include "\z\RegalisSCP_CB\addons\main\script_macros.hpp"
